requirements.txt
accelerate==1.6.0
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
attrs==25.1.0
azure-common==1.1.28
azure-core==1.33.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-storage==22.1.1
azureml-core==1.60.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
bertopic==0.16.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.0
db-dtypes==1.3.1
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud==0.34.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.22.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.2
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h2==4.2.0
hdbscan==0.8.40
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpickle==4.0.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
knack==0.12.0
langdetect==1.0.9
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.24.0
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.8
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.49
propcache==0.3.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.3
sentence-transformers==3.4.0
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
wandb==0.19.9
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.19.0
